+++ CHANGELOG.rst
@@ -39,6 +39,8 @@ Fixes
 * Prepend history of operations within dataset attributes instead of appending them. (:pull:`238`).
 * ``xsdba._processing._adapt_freq`` now avoids inserting nans. This was solved using `(dP0 <= 0)| (dP0.isnull())` and not `dP0 < 0` as a condition. (:pull:`239`, :issue:`248`, :pull:`249`).
 * ``xsdba.properties.decorrelation_length`` now handles all-nan slices correctly. (:issue:`255`, :pull:`256`).
+* Fix `adapt_freq` processing in training when ``xsdba.Grouper`` uses `add_dims`. (:issue:`182`, :pull:`183`).
+
 
 Internal changes
 ^^^^^^^^^^^^^^^^